+++ src/cross_genail.h
@@ -13,7 +13,7 @@ class GENAIL : public QTLCross
 
     GENAIL(int nf){
         crosstype = "genail";
-        phase_known_crosstype = "genail";
+        phase_known_crosstype = "genailpk"; // this actually doesn't exist
         if(nf < 2) {
             throw std::range_error("general AIL should have >= 2 founders");
         }
@@ -36,6 +36,8 @@ class GENAIL : public QTLCross
 
     const int ngen(const bool is_x_chr);
     const int nalleles();
+
+    const Rcpp::NumericMatrix geno2allele_matrix(const bool is_x_chr);
 
     const bool check_crossinfo(const Rcpp::IntegerMatrix& cross_info, const bool any_x_chr);