+++ tests/testthat/test-plot_genoprob.R
@@ -2,6 +2,8 @@ context("plot_genoprob")
 
 test_that("plot_genoprob works", {
 
+    skip_if(isnt_karl(), "plot tests only run locally")
+
     iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2"))
     iron <- iron[c("116", "232"),2]
     map <- insert_pseudomarkers(iron$gmap, step=1)
@@ -13,6 +15,6 @@ test_that("plot_genoprob works", {
         plot_genoprob(probs, map, ind="232", main="232")
     }
 
-    vdiffr::expect_doppelganger("plot_genoprob", test_plot_genoprob)
+    expect_doppelganger("plot_genoprob", test_plot_genoprob)
 
 })